app/requirements.txt
matplotlib==3.8.2
pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3
streamlit==1.39.0
urllib3==2.4.0
reportlab==4.4.1
streamlit_authenticator==0.4.2
selenium==4.22.0
webdriver-manager==4.0.2
beautifulsoup4==4.13.4
altair == 5.5.0
